Moves inspection logic into AwaitableInspector.
